Title: 10개 입력받기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231126/10개 입력받기/receive-10-inputs.c
@@ -14,7 +14,7 @@ int main() {
         sum+=arr[i];
         cnt++;
         if(cnt > 0){
-            avg = sum / cnt;
+            avg = (double)sum / cnt;
         }
     }
     printf("%d %.1lf", sum, avg);
